refactor: rename playerListerners to playerListeners, decouple voiceConnectionErrorListener
